server/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
constantly==15.1.0
coverage==6.4
cryptography==37.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
whitenoise==6.1.0
zope.interface==5.4.0
